fix(YouTube - Hide layout components): `Hide filter bar in related videos` breaks the `More videos` overlay on YouTube 20.21.37
