+++ patches/src/main/kotlin/app/morphe/patches/youtube/misc/fix/videoactionbar/RestoreOldVideoActionBarPatch.kt
@@ -21,6 +21,8 @@ import app.morphe.patches.youtube.misc.playservice.is_20_30_or_greater
 import app.morphe.patches.youtube.misc.playservice.versionCheckPatch
 import app.morphe.patches.youtube.misc.settings.PreferenceScreen
 import app.morphe.patches.youtube.misc.settings.settingsPatch
+import app.morphe.patches.youtube.shared.ModernRelateVideoOverlayFingerprint
+import app.morphe.patches.youtube.shared.RelateVideoOverlayLayoutParamFingerprint
 import app.morphe.util.getReference
 import app.morphe.util.insertLiteralOverride
 import app.morphe.util.registersUsed
@@ -108,12 +110,12 @@ internal val restoreOldVideoActionBarPatch = bytecodePatch(
                 }
             }
 
-            // If cold config data is overridden, the related video overlay breaks in fullscreen.
-            // As a workaround for this, restore the old related video overlay.
+            // fix: related video overlay is broken due to patch.
             listOf(
                 ModernRelateVideoOverlayFingerprint,
                 RelateVideoOverlayLayoutParamFingerprint
             ).forEach { fingerprint ->
+                fingerprint.clearMatch()
                 fingerprint.matchAll().forEach {
                     it.method.insertLiteralOverride(
                         it.instructionMatches.first().index,
